feat: Implment transaction processing from CSV ingestion
- Added new transaction types (Deposit, Withdrawal, Dispute, Resolve, Chargeback) to improve transaction handling.
- Implemented tracing support in CLI for better logging and debugging.
- Updated CSV source to handle different field names and types, allowing for more flexible input.
- Introduced error handling for CSV deserialization, ignoring specific errors while logging them.
- Impl'd in-memory storage to manage accounts and transactions more effectively, including support for disputed and reverted transactions.
- Added integration tests for direct transaction processing with various scenarios, ensuring robustness.
- Updated launch configuration for debugging specific test cases.
- Created a new simple test input CSV file to validate the transaction processor's functionality.

+++ crates/transaction-processor/src/sources/mod.rs
@@ -3,11 +3,11 @@ pub mod csv_source;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 
-/// Consistent-hash router: maps a `u32` key to a channel index in
+/// Consistent-hash router: maps a `u16` key to a channel index in
 /// `[0, parallelism)`. The result depends on both `key` and `parallelism`,
 /// matching the project requirement that the routing function be parameterised
 /// by the configured worker count.
-pub fn route(key: u32, parallelism: usize) -> usize {
+pub fn route(key: u16, parallelism: usize) -> usize {
     let n = parallelism.max(1);
     let mut hasher = DefaultHasher::new();
     key.hash(&mut hasher);
@@ -21,7 +21,7 @@ mod tests {
     #[test]
     fn route_is_in_range() {
         for parallelism in [1, 2, 4, 8, 16] {
-            for key in 0..1024u32 {
+            for key in 0..1024u16 {
                 assert!(route(key, parallelism) < parallelism);
             }
         }
@@ -30,6 +30,6 @@ mod tests {
     #[test]
     fn route_is_deterministic() {
         assert_eq!(route(42, 8), route(42, 8));
-        assert_eq!(route(u32::MAX, 4), route(u32::MAX, 4));
+        assert_eq!(route(u16::MAX, 4), route(u16::MAX, 4));
     }
 }
